feat: expose JSON Schema through API
